+++ src/gameboard.js
@@ -1,12 +1,7 @@
-let gameBoard = ['', '', '', '', '', '', '', '', ''];
-
 const gameBoardReset = () => {
   const squares = document.querySelectorAll('.game-board .cell');
-  Array.from(squares).forEach((e) => (e.innerText = ''));
+  Array.from(squares).forEach((e) => (e.innerText === ''));
 };
 
-const gameBoardValue = (token, position) => {
-  gameBoard[position] = token;
-};
 
-export {gameBoard, gameBoardReset, gameBoardValue}
+export default gameBoardReset;
